aichildedu/requirements.txt
# Core Framework
fastapi==0.98.0
uvicorn==0.22.0
pydantic==1.10.8
starlette==0.27.0

# Database
sqlalchemy==2.0.15
psycopg2-binary==2.9.6
pymongo==4.4.0
motor==3.2.0
redis==4.5.5
prisma==0.8.2

# Authentication and Security
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
bcrypt==4.0.1

# AI and ML
langchain==0.0.225
openai==0.27.8
tensorflow==2.12.0
torch==2.0.1
transformers==4.30.2
pillow==9.5.0
opencv-python==4.7.0.72
ffmpeg-python==0.2.0

# Utils
python-dotenv==1.0.0
requests==2.31.0
httpx==0.24.1
pydantic[email]==1.10.8
tenacity==8.2.2
pytest==7.3.1
black==23.3.0
isort==5.12.0
mypy==1.3.0

# Distributed Systems
kafka-python==2.0.2
celery==5.3.1 